tests: 100% coverage
