+++ tests/Workflow.test.ts
@@ -7,7 +7,7 @@ describe('Instantiate Workflow', () => {
   });
 
   const job = new CircleCI.Job('my-job', docker, [helloWorld]);
-  const myWorkflow = new CircleCI.workflow.Workflow('my-workflow', [job]);
+  const myWorkflow = new CircleCI.Workflow('my-workflow', [job]);
 
   const generatedWorkflow = myWorkflow.generate();
   const expected = { 'my-workflow': { jobs: ['my-job'] } };
@@ -31,7 +31,7 @@ describe('Instantiate Workflow with a custom name', () => {
     command: 'echo hello world',
   });
   const job = new CircleCI.Job('my-job', docker, [helloWorld]);
-  const myWorkflow = new CircleCI.workflow.Workflow('my-workflow');
+  const myWorkflow = new CircleCI.Workflow('my-workflow');
   myWorkflow.addJob(job, { name: 'custom-name' });
   const generatedWorkflow = myWorkflow.generate();
   const expected = {
@@ -58,7 +58,9 @@ describe('Instantiate a new Workflow with a job in the constructor', () => {
     command: 'echo hello world',
   });
   const job = new CircleCI.Job('my-job', docker, [helloWorld]);
-  const myWorkflow = new CircleCI.workflow.Workflow('my-workflow', [job]);
+  const myWorkflow = new CircleCI.Workflow('my-workflow', [
+    new CircleCI.workflow.WorkflowJob(job),
+  ]);
   const generatedWorkflow = myWorkflow.generate();
   const expected = {
     'my-workflow': { jobs: ['my-job'] },
@@ -74,7 +76,7 @@ describe('Parse a workflow', () => {
     command: 'echo hello world',
   });
   const job = new CircleCI.Job('my-job', docker, [helloWorld]);
-  const myWorkflow = new CircleCI.workflow.Workflow('my-workflow', [job]);
+  const myWorkflow = new CircleCI.Workflow('my-workflow', [job]);
 
   const workflowListShape = {
     'my-workflow': { jobs: ['my-job'] },
@@ -100,9 +102,7 @@ describe('Instantiate a new Workflow with a workflow job added manually', () => 
   });
   const job = new CircleCI.Job('my-job', docker, [helloWorld]);
   const workflowJob = new CircleCI.workflow.WorkflowJob(job);
-  const myWorkflow = new CircleCI.workflow.Workflow('my-workflow', [
-    workflowJob,
-  ]);
+  const myWorkflow = new CircleCI.Workflow('my-workflow', [workflowJob]);
 
   const generatedWorkflow = myWorkflow.generate();
   const expected = {
@@ -121,7 +121,7 @@ describe('Instantiate a new Workflow with a when condition', () => {
   const job = new CircleCI.Job('my-job', docker, [helloWorld]);
   const workflowJob = new CircleCI.workflow.WorkflowJob(job);
   const { and, or, equal, not } = CircleCI.logic;
-  const myWorkflow = new CircleCI.workflow.Workflow(
+  const myWorkflow = new CircleCI.Workflow(
     'my-workflow',
     [workflowJob],
     new CircleCI.logic.When(
@@ -173,7 +173,7 @@ describe('Utilize workflow job filters', () => {
   const job = new CircleCI.Job('my-job', docker, [helloWorld]);
 
   it('Should create branch filter', () => {
-    const myWorkflow = new CircleCI.workflow.Workflow('my-workflow');
+    const myWorkflow = new CircleCI.Workflow('my-workflow');
     myWorkflow.addJob(job, {
       filters: { branches: { only: ['/server\\/.*/'] } },
     });
@@ -185,7 +185,7 @@ describe('Utilize workflow job filters', () => {
   });
 
   it('Should create tag filter', () => {
-    const myWorkflow = new CircleCI.workflow.Workflow('my-workflow');
+    const myWorkflow = new CircleCI.Workflow('my-workflow');
     myWorkflow.addJob(job, { filters: { tags: { only: ['/^v.*/'] } } });
     const generatedWorkflowJob = myWorkflow.jobs[0].generate();
     const expected = {
@@ -203,7 +203,7 @@ describe('Utilize workflow job matrix', () => {
   const job = new CircleCI.Job('my-job', docker, [helloWorld]);
 
   it('Should create a parameter matrix', () => {
-    const myWorkflow = new CircleCI.workflow.Workflow('my-workflow');
+    const myWorkflow = new CircleCI.Workflow('my-workflow');
     myWorkflow.addJob(job, {
       matrix: {
         versions: ['1.0.0', '2.0.0'],
@@ -230,13 +230,19 @@ describe('Instantiate Workflow with a manual approval job', () => {
   });
   const jobTest = new CircleCI.Job('test-job', docker, [helloWorld]);
   const jobDeploy = new CircleCI.Job('deploy-job', docker, [helloWorld]);
+  const workflowApproval = new CircleCI.workflow.WorkflowJobApproval('on-hold');
 
-  const myWorkflow = new CircleCI.workflow.Workflow('my-workflow');
+  const myWorkflow = new CircleCI.Workflow('my-workflow', [workflowApproval]);
   myWorkflow.addJob(jobTest);
-  myWorkflow.addJobApproval('on-hold');
+  myWorkflow.addJobApproval('on-hold-2');
   myWorkflow.addJob(jobDeploy);
   const workflowContents = {
-    jobs: ['test-job', { 'on-hold': { type: 'approval' } }, 'deploy-job'],
+    jobs: [
+      { 'on-hold': { type: 'approval' } },
+      'test-job',
+      { 'on-hold-2': { type: 'approval' } },
+      'deploy-job',
+    ],
   };
   const expected = {
     'my-workflow': workflowContents,
@@ -249,14 +255,17 @@ describe('Instantiate Workflow with a manual approval job', () => {
 
   it('Should match the expected output', () => {
     expect(
-      CircleCI.parsers.parseWorkflow(
-        'my-workflow',
-        {
-          jobs: ['test-job', { 'on-hold': { type: 'approval' } }, 'deploy-job'],
-        },
-        [jobTest, jobDeploy],
-      ),
+      CircleCI.parsers.parseWorkflow('my-workflow', workflowContents, [
+        jobTest,
+        jobDeploy,
+      ]),
     ).toEqual(myWorkflow);
+  });
+
+  it('Workflow approval should be instanceof WorkflowJobAbstract', () => {
+    expect(
+      workflowApproval instanceof CircleCI.workflow.WorkflowJobAbstract,
+    ).toEqual(true);
   });
 });
 
@@ -267,7 +276,7 @@ describe('Instantiate a Workflow with sequential jobs', () => {
   });
   const jobA = new CircleCI.Job('my-job-A', docker, [helloWorld]);
   const jobB = new CircleCI.Job('my-job-B', docker, [helloWorld]);
-  const myWorkflow = new CircleCI.workflow.Workflow('my-workflow');
+  const myWorkflow = new CircleCI.Workflow('my-workflow');
   myWorkflow.addJob(jobA);
   myWorkflow.addJob(jobB, { requires: ['my-job-A'] });
   it('Should match the expected output', () => {
@@ -286,7 +295,7 @@ describe('Instantiate a Workflow with 2 jobs', () => {
   const helloWorld = new CircleCI.commands.Run({ command: 'echo hello world' });
   const jobA = new CircleCI.Job('my-job-A', docker, [helloWorld]);
   const jobB = new CircleCI.Job('my-job-B', docker, [helloWorld]);
-  const myWorkflow = new CircleCI.workflow.Workflow('my-workflow');
+  const myWorkflow = new CircleCI.Workflow('my-workflow');
   myWorkflow.addJob(jobA, { myParam: 'my-value' });
   myWorkflow.addJob(jobB);
   it('Should match the expected output', () => {
@@ -312,7 +321,7 @@ describe('Add a job to a workflow with a custom name parameter', () => {
     command: 'echo hello world',
   });
   const job = new CircleCI.Job('my-job', docker, [helloWorld]);
-  const myWorkflow = new CircleCI.workflow.Workflow('my-workflow');
+  const myWorkflow = new CircleCI.Workflow('my-workflow');
   myWorkflow.addJob(job, { name: 'custom-name' });
   it('Should match the expected output', () => {
     const expected = {
